Allows providing headers to PartiQLPlanner
