+++ partiql-planner/src/main/kotlin/org/partiql/planner/PartiQLPlannerBuilder.kt
@@ -7,10 +7,11 @@ import org.partiql.spi.Plugin
  */
 class PartiQLPlannerBuilder {
 
+    private var headers: MutableList<Header> = mutableListOf(PartiQLHeader)
     private var plugins: List<Plugin> = emptyList()
     private var passes: List<PartiQLPlannerPass> = emptyList()
 
-    fun build(): PartiQLPlanner = PartiQLPlannerDefault(plugins, passes)
+    fun build(): PartiQLPlanner = PartiQLPlannerDefault(headers, plugins, passes)
 
     public fun plugins(plugins: List<Plugin>): PartiQLPlannerBuilder = this.apply {
         this.plugins = plugins
@@ -18,5 +19,9 @@ class PartiQLPlannerBuilder {
 
     public fun passes(passes: List<PartiQLPlannerPass>): PartiQLPlannerBuilder = this.apply {
         this.passes = passes
+    }
+
+    public fun headers(headers: List<Header>): PartiQLPlannerBuilder = this.apply {
+        this.headers += headers
     }
 }
